+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PixPullSubscription resource
 - PixPullRequest resource
 - IssuingStockRule resource
+- PixFraud resource
+- PixKeyHolmes resource
+- PixInternalTransactionReport resource
 ### Changed
 - minimum required Ruby version to 3.0
 ### Fixed